refactor: #295 애플로그인 테스트 로직 request로 합침
