--- ByeBoo-iOS/ByeBoo-iOS/Data/Network/Service/AppleToken.swift
@@ -1,20 +1,0 @@
-//
-//  AppleToken.swift
-//  ByeBoo-iOS
-//
-//  Created by 이나연 on 10/31/25.
-//
-
-struct AppleToken {
-    let identityToken: String
-    let authorizationCode: String
-}
-
-extension AppleToken {
-    static func stub() -> AppleToken {
-        .init(
-            identityToken: "identityToken",
-            authorizationCode: "authorizationCode"
-        )
-    }
-}
